3_sem/Time/finance_prediction_bot/requirements.txt
python-telegram-bot==20.7
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
scikit-learn==1.4.2
statsmodels==0.14.1
tensorflow==2.16.2
pmdarima==2.0.4